refactor: reducer 분리
