+++ src/redux/store.js
@@ -2,9 +2,9 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { persistStore } from 'redux-persist';
-import rootReducer from './reducer';
+import persistedReducer from './reducers';
 
-const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
+const store = createStore(persistedReducer, composeWithDevTools(applyMiddleware(thunk)));
 
 const persistedStore = persistStore(store);
